fix(mcp): satisfy three linters CI saw and a cached local run did not
`make check-backend` passed locally and `deterministic-gates` failed on the same
commit, because golangci-lint's analysis cache had the previous edit's results. The
findings were real:

- **goconst** — `"uuid"` appeared three times in the schema derivation. Now
  `uuidFormat`, which is worth naming beyond the linter: the derivation matches that
  format at three levels of the schema, and a typo in one would leave that level
  silently unguarded.
- **nilerr** — two branches returned nil after a non-nil unmarshal error. The intent
  was "this shape is not mine to judge; decodeArgs says it better", and the linter
  was right that the code did not say so. Both now go through `argsAsObject` /
  `itemsAsObjects`, which return a shape verdict instead of swallowing an error.
- **staticcheck ST1020** — a doc comment on an exported `Error` must begin "Error".

`registry.go` then crossed the 500-LOC cap, so the id-claim concept moved to
`idargs.go`: registry.go decides whether a call may run at all, idargs.go decides
whether its ids name anything.

Worth remembering: a green local `make check-backend` does not prove lint is clean
after an edit, because the lint cache is keyed on inputs golangci-lint has already
seen. CI runs cold.
